Agent_Dhruv/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
sentence-transformers==3.0.1
transformers>=4.41.0
openai>=1.35.0
tiktoken>=0.7.0
faiss-cpu>=1.7.4
python-multipart>=0.0.9
